refactored phone scenario
