+++ package.json
@@ -53,10 +53,6 @@
 				"src": "./src/popup-content.js"
 			},
 			{
-				"name": "phone-utils",
-				"src": "./src/phone-utils.js"
-			},
-			{
 				"name": "material-utils",
 				"src": "./src/material-utils.js"
 			},
