+++ src/autodetect/freight/freight-ui.js
