+++ src/manifest.json
@@ -25,7 +25,7 @@
     ],
     "content_scripts": [
       {
-        "js": ["content.js", "popup-content.js", "phone-utils.js", "material-utils.js"],
+        "js": ["content.js", "popup-content.js", "material-utils.js"],
         "matches": ["<all_urls>"]
       }
     ],
